# Nmer Generation Speedup