## Description
This PR speeds up the generation of nmers in CrystaLatte. This is done by:
 - Replacing list iterations with efficient numpy functions in the calculations of distance matrices, chemical space eigenvalues, and the nuclear repulsion energy
 - Changing the psithon file writer from writing one character at a time (many `os.write()` calls per input file) to a single write (one `os.write()` call per input file).
 - Improving the duplication check for potential new nmers. Previously, a potential new nmer was compared to every single existing nmer to check for duplication. Now, a potential new nmer is only compared with existing nmers with similar nuclear repulsion energies (+- 1e-5). This is accomplished by maintaining a list of existing nmers sorted by nuclear repulsion energy.

I tested this PR for correctness and efficiency using the following `.cle` input:
```
cif_input       = Benzene.cif
cif_output      = Benzene.xyz
cif_a           = 9 
cif_b           = 9 
cif_c           = 9 
nmers_up_to     = 3 
r_cut_com       = 20.0
r_cut_monomer   = 20.0
r_cut_dimer     = 18.0
r_cut_trimer    = 12.0
cle_run_type    = psithon
psi4_method     = HF/STO-3G
psi4_bsse       = nocp
psi4_memory     = 500 MB
verbose         = 2 
```
For this test, the number of dimers (88) and trimers (689) was unchanged. Also, the identities of the dimers and trimers are unchanged. I checked this by comparing the table of nmers printed by CrystaLatte. 

For this test, this PR reduces the total wall time from 903 seconds to 103 seconds (a ~9x speedup). Some other observations:
 - The BFS algorithm takes 38 seconds for this input. The remaining time is spent generating dimers and trimers
 - The dimer generation is sped up from 2.0 seconds to 0.4 seconds (a ~5x speedup)
 - The trimer generation is sped up from 862 seconds to 63 seconds (a ~14x speedup)

I expect the speedup factor (~9x here) to grow for larger supercells.

## TODO
- [x] Add additional unit tests

## Status
- [x] Ready to go