@carlosborca 2021-07-09
Alright, could you please add the aforementioned tests for the modified functions in the _N_-mers generation routine? That would also be greatly appreciated.